requirements.txt
pandas==2.2.3
openpyxl==3.1.5
selenium==4.26.0
requests==2.32.3
backoff==2.2.1
beautifulsoup4==4.12.3
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
oauth2client==4.1.3
pylint==3.3.2
gunicorn==23.0.0
PyPDF2==3.0.1
google-analytics-data==0.18.14
jsonlines==4.0.0
joblib==1.4.2
python-dotenv==1.0.1
playwright==1.48.0
fastapi[standard]==0.115.4
boto3==1.35.60
pika==1.3.2
opentelemetry-api==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-exporter-otlp-proto-grpc==1.28.1
opentelemetry-instrumentation-fastapi==0.49b1
opentelemetry-instrumentation-logging==0.49b1
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2
psycopg==3.2.3
asyncpg==0.30.0
greenlet==3.1.1
